update height message from JS comes in as Double

+++ ProtonMail/ProtonMailCommon/Utilities/WebContentsSecureLoader/HTTPRequestSecureLoader.swift
@@ -188,11 +188,11 @@ class HTTPRequestSecureLoader: NSObject, WebContentsSecureLoader, WKScriptMessag
             
             self.webView?.load(request)
         }
-        if let preheight = dict["preheight"] as? CGFloat {
-            self.renderedContents.preheight = preheight
+        if let preheight = dict["preheight"] as? Double {
+            self.renderedContents.preheight = CGFloat(preheight)
         }
-        if let height = dict["height"] as? CGFloat {
-            self.renderedContents.height = height
+        if let height = dict["height"] as? Double {
+            self.renderedContents.height = CGFloat(height)
         }
     }
